OM-8 - add-unique-data-to-dashboard
description: added aerospike unique-data utilization of memory, disk and pmem in a cluster.
as part o this, modified the Cluster view.json and below are the changes made
1. added a new row "Unique Data Usage" below Overview row
2. in "Unique Data Usage" row, added 3 new panels
2.A - Panel 1: Displays Cumulative "Master Objects" and "Unique Bytes" across the cluster
2.B - Panel 2: Namespace wise "Master Objects" - displays "master objects" in each namespace
2.C - Pane 3: Namespace wise - "Unique Data Utilization"

below is the confluence location where this Unique Data usage formula is documents and this data is validated against the UDA Tool.

Confluence page:
https://aerospike.atlassian.net/wiki/spaces/PRODUCT/pages/198344706/License+Data+Formulae