OM-8 - add-unique-data-to-dashboard
2 key changes in this check-in:
removed the Unique-Data usage row and panel from the cluster view dashboard, as this unique data is not matching with the philosophy of cluster view monitoring aspects.

added a new dashboard called uniquedata.json which display all the data elements related to unique data in a cluster. this has 2 data display models.
1. cumulative sum of all master objects and unique data bytes across all the nodes in the selected cluster
2. display a panel for each namespace with namespace specific master objects and unique data bytes.